Fixes to the Loop handling and more
Make the loop variable be either a Python int or float. We don't want
a pure integer loop variable, which is what is typically used, to
expand with a pointless ".0" at the end.

If a Loop element is in a template expansion, don't make a copy of the
params dict as the very point of using the Loop element is to have
Parameters elements inside it that modify the parameters for the
invocation of the enclosing template. Real MSFS XML programs use loops
and conditions only inside templates, so modify the "unit test"
loop.xml to test Loop both inside templates and outside. condition.xml
already tests the Condition element inside a template.

Other miscellaneous changes.